rename pin.h to nrf_pin.h to avoid conflict with Pin.h
